Move LoggingInvocationHandler to presto-plugin-toolkit
This is useful class, that could also be used in other connectors.
