+++ presto-thrift-testing-server/pom.xml
@@ -26,6 +26,12 @@
         <dependency>
             <groupId>io.airlift.drift</groupId>
             <artifactId>drift-server</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>io.airlift</groupId>
+                    <artifactId>parameternames</artifactId>
+                </exclusion>
+            </exclusions>
         </dependency>
 
         <dependency>
